refactor: 프로젝트 구조 개선 및 모듈 분리 - core/ 디렉토리 추가로 핵심 시스템 모듈화 - models/ 디렉토리 생성 (데이터 모델 준비) - services/ 리팩터링 (problem_service.py로 이전) - utils/ 확장 (validators, helpers 모듈 준비) - import 경로 업데이트 (config → core.config) - 불필요한 예제 파일 제거
